Move PSP announcement resource into domain package
