+++ internal/provider/pspannouncement/resource.go
@@ -1,4 +1,4 @@
-package provider
+package pspannouncement
 
 import (
 	"context"
@@ -19,6 +19,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/client"
+	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/providerclient"
 )
 
 var (
@@ -28,8 +29,8 @@ var (
 	_ resource.ResourceWithValidateConfig = &pspAnnouncementResource{}
 )
 
-// NewPSPAnnouncementResource returns the PSP announcement resource implementation.
-func NewPSPAnnouncementResource() resource.Resource {
+// NewResource returns the PSP announcement resource implementation.
+func NewResource() resource.Resource {
 	return &pspAnnouncementResource{}
 }
 
@@ -60,20 +61,7 @@ type pspAnnouncementExpected struct {
 }
 
 func (r *pspAnnouncementResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
-	if req.ProviderData == nil {
-		return
-	}
-
-	client, ok := req.ProviderData.(*client.Client)
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			"The provider data is not of type *client.Client",
-		)
-		return
-	}
-
-	r.client = client
+	r.client = providerclient.FromResourceConfigure(req, resp)
 }
 
 func (r *pspAnnouncementResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
