+++ internal/provider/pspannouncement/resource_test.go
@@ -1,4 +1,4 @@
-package provider
+package pspannouncement
 
 import (
 	"encoding/json"
